feat: extend dataset and framework compatibility

+++ starVLA/model/framework/VLM4A/QwenPI.py
@@ -283,7 +283,7 @@ class Qwen_PI(baseframework):
                 vl_embs_list, state
             )  # (B, chunk_len, action_dim)
 
-        normalized_actions = pred_actions.detach().cpu().numpy()
+        normalized_actions = pred_actions.detach().float().cpu().numpy()
         return {"normalized_actions": normalized_actions}
